use big decimal for storing balance

+++ src/models/balance_pack.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use serde::{Deserialize, Serialize};
 
 use super::wallet::WalletC;
@@ -6,13 +7,13 @@ use super::wallet::WalletC;
 #[derive(Debug, Serialize, Deserialize)]
 pub struct BalancePack {
     pub server_http_address:String,
-    pub balance:f32
+    pub balance:BigDecimal
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct BalanceCPack {
     pub ip_address:String,
-    pub balance:f32
+    pub balance:BigDecimal
 }
